gradio_app/requirements.txt
pandas==1.3.5
pandas-datareader==0.9.0
pandas-gbq==0.17.9
pandas-profiling==1.4.1
sklearn-pandas==1.8.0
numpy==1.21.6